prediction: split beam_search_parlays into parlay.py (Phase 3 Step 3 / session 5)
CONTRIBUTING.md §Package Map advertised that beam_search_parlays lives in
prediction/parlay.py while find_correlation stays in correlation.py. Until
now both lived in correlation.py. Carve out the new module so the package
matches the contract.

Moved to prediction/parlay.py:
  - beam_search_parlays
  - _payout_curve_for, _expected_payout_with_pushes, _nearest_psd
  - all beam-search constants (_BEAM_WIDTH, _PARLAY_GEO_MEAN_FLOOR,
    _MIN_PRODUCT_BOOST, _BOOKS_EV_FLOOR, _MODEL_EV_PRECHECK_FLOOR,
    _MODEL_EV_FINAL_FLOOR, _KELLY_UNITS_FLOOR, _PSD_EIG_TOLERANCE,
    _PAYOUT_CLIP_LO/HI, _PUSH_PROB_FLOOR, _PUSH_MC_SAMPLES,
    _KELLY_BANKROLL_FRACTION)

Stayed in prediction/correlation.py:
  - find_correlation, _team_slice, _build_game_corr_map
  - _legacy_underdog_overwrite_payouts / _legacy_underdog_search_payouts
    (deferred-imported by parlay._payout_curve_for to avoid a cycle)
  - _MAX_BOOST_UNDERDOG / _MAX_BOOST_OTHER (only used by find_correlation)
  - back-compat re-exports of the moved helpers so existing tests and
    callers can keep importing from prediction.correlation

prediction/__init__.py re-exports beam_search_parlays at package scope.
Behavior is unchanged; this is a pure file-split.

Tests: new tests/golden/test_prediction_imports.py pins both the canonical
parlay-module path and the package-level re-export.
